requirements.txt
PyGithub>=1.55
matplotlib>=3.4
numpy>=1.20